+++ magda-web-client/src/Components/Dataset/Add/DatasetAddMetadataPage.tsx
@@ -47,7 +47,10 @@ import urijs from "urijs";
 import FileDeletionError from "helpers/FileDeletionError";
 import redirect from "helpers/redirect";
 import Loader from "rsuite/Loader";
-import inPopUpMode from "helpers/inPopUpMode";
+import {
+    createPopupModeQueryString,
+    executePopupModeCallback
+} from "helpers/popupUtils";
 
 type Props = {
     initialState: State;
@@ -68,7 +71,7 @@ type Props = {
 class NewDataset extends React.Component<Props, State> {
     state: State = this.props.initialState;
 
-    isInPopUpMode = inPopUpMode();
+    popUpModeQueryString = createPopupModeQueryString();
 
     constructor(props) {
         super(props);
@@ -308,7 +311,13 @@ class NewDataset extends React.Component<Props, State> {
                 redirect(this.props.history, `/dataset/list`);
             } else {
                 // redirect to datasets management
-                redirect(this.props.history, `/settings/datasets/draft`);
+                redirect(
+                    this.props.history,
+                    `/settings/datasets/draft` +
+                        (this.popUpModeQueryString
+                            ? `?${this.popUpModeQueryString}`
+                            : "")
+                );
             }
         } catch (e) {
             this.props.createNewDatasetError(e);
@@ -338,7 +347,9 @@ class NewDataset extends React.Component<Props, State> {
                         encodeURIComponent(this.props.datasetId) +
                         "/" +
                         step +
-                        (this.isInPopUpMode ? `?popup=true` : "")
+                        (this.popUpModeQueryString
+                            ? `?${this.popUpModeQueryString}`
+                            : "")
                 );
             }
         } catch (e) {
@@ -407,10 +418,13 @@ class NewDataset extends React.Component<Props, State> {
             if (result.length) {
                 throw new FileDeletionError(result);
             }
+            executePopupModeCallback(datasetId);
             redirect(
                 this.props.history,
                 `/dataset/add/metadata/${encodeURIComponent(datasetId)}/6` +
-                    (this.isInPopUpMode ? `?popup=true` : "")
+                    (this.popUpModeQueryString
+                        ? `?${this.popUpModeQueryString}`
+                        : "")
             );
         } catch (e) {
             this.setState({