+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## v5.3.0
 
 - New feature: Access Group UI, useful when you need arbitrary access control management of datasets. Related to: #3269
+- #3600: Allow dataset editor tool UI notify popup launcher when the dataset has been submitted
 
 ## v5.2.0